refactor : 패키지 정리, 2개의 lotteryticket 겹치는 수 구하는 것 도메인 객체의 메서드로 분리
