+++ src/test/java/lotto/model/result/LotteryGameResultDtoTest.java
@@ -1,5 +1,7 @@
-package lotto.model.domain;
+package lotto.model.result;
 
+import lotto.model.domain.Rank;
+import lotto.model.result.LotteryGameResultDto;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
